create directory to recreate results from oral exam.
I fixed the function that outputs the number of matsubara frequencies
to be consistent with that of the code from irb.py on the supercomputer.
However, the lambda bar values that are being spit out of each
calculation(this code vs the supercomputer code) are not identical.
This must be fixed.